feat(business): implement dialog component
Extract components and directives into separate files
