+++ projects/sbb-esta/angular-business/dialog/src/public_api.ts
@@ -1,7 +1,10 @@
 export * from './dialog.module';
 export * from './dialog/dialog.service';
 export * from './dialog/dialog-container.component';
-export * from './dialog/dialog-content';
+export * from './dialog/dialog.component';
+export * from './dialog/dialog-header.component';
+export * from './dialog/dialog-content.component';
+export * from './dialog/dialog-footer.component';
 export * from './dialog/dialog-config';
 export * from './dialog/dialog-ref';
 export * from './dialog/dialog-animations';
